Implemented solution with application
