+++ exercises/Calendar/README.md
@@ -1,2 +1,5 @@
-Este es el ejercicio 3 de la práctica 1 el cual permite almacenar los eventos que introduzca el usuario en un calendario del año 2021 e imprime los eventos que 
-el usuario haya ingresado con el formato "El evento < nombre_evento > será el dia < dia_de_la_semana_evento >(Ej: Viernes) < dia >(Ej:7) de < mes > del año 2021"
+# Calendar
+
+## UCV. Semestre 1-2021. Práctica 1. Ejercicio 3
+
+Defina una estructura de datos que permita almacenar eventos de un calendario. Ésta no debe registrar eventos de fechas inexistentes. Asuma que su calendario funciona para un solo año. Además se desea conocer el día de la semana en que ocurre el evento.
