--- exercises/exercise3.1-Calendar/README.md
@@ -1,5 +1,0 @@
-# Calendar
-
-## UCV. Semestre 1-2021. Práctica 1. Ejercicio 3
-
-Defina una estructura de datos que permita almacenar eventos de un calendario. Ésta no debe registrar eventos de fechas inexistentes. Asuma que su calendario funciona para un solo año. Además se desea conocer el día de la semana en que ocurre el evento.
